+++ src/extension.ts
@@ -19,23 +19,27 @@ export function activate(context: vscode.ExtensionContext) {
     const editor = vscode.window.activeTextEditor;
     if (!editor) {
       return;
-    }
+    }    
     const code = getCode(editor)
 
     const document = editor.document;
     const languageId = document.languageId;
     try {
-      const result = await executeCode(code, languageId);
+      const config = vscode.workspace.getConfiguration('venkat');
+      const resultAsComment = config.get<boolean>('resultAsComment')===true;
+
+      const result = await executeCode(code, languageId, resultAsComment);
 
       if (result) {
-        const position = document.lineAt(editor.selection.end).range.end;
+        if (resultAsComment) {          
+          const position = document.lineAt(editor.selection.end).range.end;
+          editor.edit((editBuilder) => editBuilder.insert(position, result));
 
-        editor.edit((editBuilder) => {
-          editBuilder.insert(position, result);
-        });
-
-        const newSelection = new vscode.Selection(position, position.translate(0, (result).length));
-        editor.selection = newSelection;
+          const newSelection = new vscode.Selection(position, position.translate(0, (result).length));
+          editor.selection = newSelection;
+        } else {
+          vscode.window.showInformationMessage(result);
+        }
       }
     } catch (err) {
         const message : string = (err instanceof Error) ? err.message : new String(err).toString();